State ownership, GameOver state, consistent data types

+++ include/common.h
@@ -2,5 +2,5 @@
 
 #include <cstddef>
 
-template <typename T, size_t N>
-constexpr size_t arraySize(const T (&)[N]) { return N; }
+template <typename T, uint16_t N>
+constexpr uint16_t arraySize(const T (&)[N]) { return N; }
